tools/fctl/requirements.txt
docker>=6.1.3
kubernetes>=26.1.0
PyYAML==6.0
rich==13.4.2
typer==0.9.0
typing_extensions==4.7.1